[#25] Move bus stops adapter to its package
