+++ app/src/main/java/es/craftsmanship/toledo/katangapp/adapters/BusStopsAdapter.java
@@ -1,5 +1,8 @@
-package es.craftsmanship.toledo.katangapp.activities;
+package es.craftsmanship.toledo.katangapp.adapters;
 
+import es.craftsmanship.toledo.katangapp.activities.BusStopActivity;
+import es.craftsmanship.toledo.katangapp.activities.BusStopRoutesAdapter;
+import es.craftsmanship.toledo.katangapp.activities.R;
 import es.craftsmanship.toledo.katangapp.models.BusStop;
 import es.craftsmanship.toledo.katangapp.models.BusStopResult;
 import es.craftsmanship.toledo.katangapp.models.RouteResult;
